requirements.txt
google-adk>=0.3.0
google-genai>=1.0.0
yfinance>=0.2.36
pandas>=2.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
